+++ src/widget/analog-scale/analog-scalepicker.js
@@ -2,6 +2,7 @@
 var Widget = require( '../../js/Widget' );
 var $ = require( 'jquery' );
 var pluginName = 'analogscalepicker';
+var support = require( '../../js/support' );
 
 require( 'bootstrap-slider-basic' );
 
@@ -69,7 +70,7 @@ Analogscalepicker.prototype._getProps = function( $question ) {
     var type = this.element.attributes[ 'data-type-xml' ].value;
 
     return {
-        touch: this.options.touch,
+        touch: support.touch,
         readonly: this.element.readOnly,
         step: type === 'decimal' ? 0.1 : 1,
         orientation: appearances.indexOf( 'horizontal' ) !== -1 ? 'horizontal' : 'vertical'