+++ src/widget/datetime/datetimepicker-extended.js
@@ -13,7 +13,7 @@ require( '../../js/dropdown.jquery' );
  *
  * @constructor
  * @param {Element}                       element   Element to apply widget to.
- * @param {(boolean|{touch: boolean})}    options   options
+ * @param {*}    options   options
  * @param {*=}                            event     event
  */
 
@@ -172,9 +172,9 @@ $.fn[ pluginName ] = function( options, event ) {
     options = options || {};
 
     return this.each( function() {
-        var $this = $( this ),
-            data = $this.data( pluginName ),
-            badSamsung = /GT-P31[0-9]{2}.+AppleWebKit\/534\.30/;
+        var $this = $( this );
+        var data = $this.data( pluginName );
+        var badSamsung = /GT-P31[0-9]{2}.+AppleWebKit\/534\.30/;
 
         /*
             Samsung mobile browser (called "Internet") has a weird bug that appears sometimes (?) when an input field
@@ -186,7 +186,7 @@ $.fn[ pluginName ] = function( options, event ) {
             browser: "Mozilla/5.0 (Linux; U; Android 4.1.1; en-us; GT-P3113 Build/JRO03C) AppleWebKit/534.30 (KHTML, like Gecko) Version/4.0 Safari/534.30";
             webview: "Mozilla/5.0 (Linux; U; Android 4.1.2; en-us; GT-P3100 Build/JZO54K) AppleWebKit/534.30 (KHTML, like Gecko) Version/4.0 Safari/534.30" 
             */
-        if ( !data && typeof options === 'object' && ( !options.touch || !support.inputtypes.datetime || badSamsung.test( navigator.userAgent ) ) ) {
+        if ( !data && typeof options === 'object' && ( !support.touch || !support.inputTypes.datetime || badSamsung.test( navigator.userAgent ) ) ) {
             $this.data( pluginName, new DatetimepickerExtended( this, options, event ) );
         }
         //only call method if widget was instantiated before