+++ src/widget/select-mobile/selectpicker.js
@@ -2,6 +2,7 @@
 
 var $ = require( 'jquery' );
 var Widget = require( '../../js/Widget' );
+var support = require( '../../js/support' );
 var pluginName = 'mobileSelectpicker';
 
 /**
@@ -73,7 +74,7 @@ $.fn[ pluginName ] = function( options, event ) {
             data = $this.data( pluginName );
 
         //only instantiate if options is an object AND if options.touch is truthy
-        if ( !data && typeof options === 'object' && options.touch ) {
+        if ( !data && typeof options === 'object' && support.touch ) {
             $this.data( pluginName, new MobileSelectpicker( this, options, event ) );
         }
         if ( data && typeof options === 'string' ) {